Remove the use of 'uniqueIdentifier'.  Fully deprecated in iOS 7 and Xcode 5.Fix warnings associated with 64-bit/32-bit compiler.

+++ BPXLUUIDHandler.m
@@ -148,7 +148,7 @@ static CFMutableDictionaryRef CreateKeychainQueryDictionary(void)
 	
 	if (status != noErr)
 	{
-		NSLog(@"BPXLUUIDHandler Keychain Save Error: %ld", status);
+		NSLog(@"BPXLUUIDHandler Keychain Save Error: %d", (int)status);
 		uuid = nil;
 	}
 	
@@ -182,7 +182,7 @@ static NSString *_uuid = nil;
 		}
 		else // Any other error, log it and return nil
 		{
-			NSLog(@"BPXLUUIDHandler Unhandled Keychain Error %ld", status);
+			NSLog(@"BPXLUUIDHandler Unhandled Keychain Error %d", (int)status);
 			return nil;
 		}
 	}
@@ -202,7 +202,7 @@ static NSString *_uuid = nil;
 		}
 		else // Any other error, log it and return nil
 		{
-			NSLog(@"BPXLUUIDHandler Unhandled Keychain Error %ld", status);
+			NSLog(@"BPXLUUIDHandler Unhandled Keychain Error %d", (int)status);
 			return nil;
 		}
 	}
@@ -246,7 +246,7 @@ static NSString *_uuid = nil;
 	status = SecItemDelete(query);
 	if (status != noErr)
 	{
-		NSLog(@"BPXLUUIDHandler Keychain Delete Error: %ld", status);
+		NSLog(@"BPXLUUIDHandler Keychain Delete Error: %d", (int)status);
 	}
 	CFRelease(query);
 }
